b2: fixed regression when uploading zero-length blobs introduced by #1589 (#1593)
Turns out B2 library will treat seekable readers in a special way
and will pass those directly to http.NewRequest() for upload. Other
readers are copied to a temporary buffer first.

The #1589 made the reader io.ReadSeeker so it was passed to http.NewRequest(),
which for payloads of length zero uses heuristics to determine if the
Content-Length should be passed as zero or not passed at all.

https://cs.opensource.google/go/go/+/refs/tags/go1.17.5:src/net/http/request.go;l=890;drc=e6dda19888180c5159460486d30c0412e4980748

Since reader was not one of the magic types, Content-Length was not
passed at all, causing B2 server to choke on zero-length files.