docs(site): Revised main homepage (#2131)
* docs(site): Revised main homepage

* Update _index.html

* Update _index.html

* Add files via upload

We do not have to use this background image. But if we do, we should find some place to credit https://unsplash.com/photos/oyXis2kALVg. The image is free for us to use and we don't need to credit per their license -- but it is the right thing to do.

* Update _index.html

* Add files via upload

Just trying other backgrounds to see how they look.

* Add files via upload

* Add files via upload

* Update _index.html

* Update _index.html

* Update footer.html

* Update footer.html

* Update _index.html

* Update _index.html